fix: duplication in sar unstake
